Refactor Visualization Inheritance
