+++ inference/core/workflows/core_steps/visualizations/corner.py
@@ -3,9 +3,10 @@ from typing import List, Literal, Optional, Type, Union
 import supervision as sv
 from pydantic import ConfigDict, Field
 
-from inference.core.workflows.core_steps.visualizations.base import (
-    VisualizationBlock,
-    VisualizationManifest,
+from inference.core.workflows.core_steps.visualizations.base import OUTPUT_IMAGE_KEY
+from inference.core.workflows.core_steps.visualizations.base_colorable import (
+    ColorableVisualizationBlock,
+    ColorableVisualizationManifest,
 )
 from inference.core.workflows.entities.base import WorkflowImageData
 from inference.core.workflows.entities.types import (
@@ -13,8 +14,6 @@ from inference.core.workflows.entities.types import (
     WorkflowParameterSelector,
 )
 from inference.core.workflows.prototypes.block import BlockResult, WorkflowBlockManifest
-
-OUTPUT_IMAGE_KEY: str = "image"
 
 TYPE: str = "CornerVisualization"
 SHORT_DESCRIPTION = "Draws the corners of detected objects in an image."
@@ -24,7 +23,7 @@ objects in an image using Supervision's `sv.BoxCornerAnnotator`.
 """
 
 
-class CornerManifest(VisualizationManifest):
+class CornerManifest(ColorableVisualizationManifest):
     type: Literal[f"{TYPE}"]
     model_config = ConfigDict(
         json_schema_extra={
@@ -48,7 +47,7 @@ class CornerManifest(VisualizationManifest):
     )
 
 
-class CornerVisualizationBlock(VisualizationBlock):
+class CornerVisualizationBlock(ColorableVisualizationBlock):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.annotatorCache = {}
