+++ inference/core/workflows/core_steps/visualizations/dot.py
@@ -3,9 +3,10 @@ from typing import List, Literal, Optional, Type, Union
 import supervision as sv
 from pydantic import ConfigDict, Field
 
-from inference.core.workflows.core_steps.visualizations.base import (
-    VisualizationBlock,
-    VisualizationManifest,
+from inference.core.workflows.core_steps.visualizations.base import OUTPUT_IMAGE_KEY
+from inference.core.workflows.core_steps.visualizations.base_colorable import (
+    ColorableVisualizationBlock,
+    ColorableVisualizationManifest,
 )
 from inference.core.workflows.entities.base import WorkflowImageData
 from inference.core.workflows.entities.types import (
@@ -14,8 +15,6 @@ from inference.core.workflows.entities.types import (
     WorkflowParameterSelector,
 )
 from inference.core.workflows.prototypes.block import BlockResult, WorkflowBlockManifest
-
-OUTPUT_IMAGE_KEY: str = "image"
 
 TYPE: str = "DotVisualization"
 SHORT_DESCRIPTION = (
@@ -27,7 +26,7 @@ based on provided detections using Supervision's `sv.DotAnnotator`.
 """
 
 
-class DotManifest(VisualizationManifest):
+class DotManifest(ColorableVisualizationManifest):
     type: Literal[f"{TYPE}"]
     model_config = ConfigDict(
         json_schema_extra={
@@ -71,7 +70,7 @@ class DotManifest(VisualizationManifest):
     )
 
 
-class DotVisualizationBlock(VisualizationBlock):
+class DotVisualizationBlock(ColorableVisualizationBlock):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.annotatorCache = {}
