+++ inference/core/workflows/core_steps/visualizations/ellipse.py
@@ -3,9 +3,10 @@ from typing import List, Literal, Optional, Type, Union
 import supervision as sv
 from pydantic import ConfigDict, Field
 
-from inference.core.workflows.core_steps.visualizations.base import (
-    VisualizationBlock,
-    VisualizationManifest,
+from inference.core.workflows.core_steps.visualizations.base import OUTPUT_IMAGE_KEY
+from inference.core.workflows.core_steps.visualizations.base_colorable import (
+    ColorableVisualizationBlock,
+    ColorableVisualizationManifest,
 )
 from inference.core.workflows.entities.base import WorkflowImageData
 from inference.core.workflows.entities.types import (
@@ -13,8 +14,6 @@ from inference.core.workflows.entities.types import (
     WorkflowParameterSelector,
 )
 from inference.core.workflows.prototypes.block import BlockResult, WorkflowBlockManifest
-
-OUTPUT_IMAGE_KEY: str = "image"
 
 TYPE: str = "EllipseVisualization"
 SHORT_DESCRIPTION = "Draws ellipses that highlight detected objects in an image."
@@ -24,7 +23,7 @@ objects in an image using Supervision's `sv.EllipseAnnotator`.
 """
 
 
-class EllipseManifest(VisualizationManifest):
+class EllipseManifest(ColorableVisualizationManifest):
     type: Literal[f"{TYPE}"]
     model_config = ConfigDict(
         json_schema_extra={
@@ -54,7 +53,7 @@ class EllipseManifest(VisualizationManifest):
     )
 
 
-class EllipseVisualizationBlock(VisualizationBlock):
+class EllipseVisualizationBlock(ColorableVisualizationBlock):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.annotatorCache = {}
