+++ inference/core/workflows/core_steps/visualizations/label.py
@@ -3,9 +3,10 @@ from typing import List, Literal, Optional, Type, Union
 import supervision as sv
 from pydantic import ConfigDict, Field
 
-from inference.core.workflows.core_steps.visualizations.base import (
-    VisualizationBlock,
-    VisualizationManifest,
+from inference.core.workflows.core_steps.visualizations.base import OUTPUT_IMAGE_KEY
+from inference.core.workflows.core_steps.visualizations.base_colorable import (
+    ColorableVisualizationBlock,
+    ColorableVisualizationManifest,
 )
 from inference.core.workflows.core_steps.visualizations.utils import str_to_color
 from inference.core.workflows.entities.base import WorkflowImageData
@@ -17,8 +18,6 @@ from inference.core.workflows.entities.types import (
 )
 from inference.core.workflows.prototypes.block import BlockResult, WorkflowBlockManifest
 
-OUTPUT_IMAGE_KEY: str = "image"
-
 TYPE: str = "LabelVisualization"
 SHORT_DESCRIPTION = (
     "Draws labels on an image at specific coordinates based on provided detections."
@@ -29,7 +28,7 @@ based on provided detections using Supervision's `sv.LabelAnnotator`.
 """
 
 
-class LabelManifest(VisualizationManifest):
+class LabelManifest(ColorableVisualizationManifest):
     type: Literal[f"{TYPE}"]
     model_config = ConfigDict(
         json_schema_extra={
@@ -102,7 +101,7 @@ class LabelManifest(VisualizationManifest):
     )
 
 
-class LabelVisualizationBlock(VisualizationBlock):
+class LabelVisualizationBlock(ColorableVisualizationBlock):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.annotatorCache = {}
