+++ inference/core/workflows/core_steps/visualizations/mask.py
@@ -3,9 +3,10 @@ from typing import List, Literal, Optional, Type, Union
 import supervision as sv
 from pydantic import ConfigDict, Field
 
-from inference.core.workflows.core_steps.visualizations.base import (
-    VisualizationBlock,
-    VisualizationManifest,
+from inference.core.workflows.core_steps.visualizations.base import OUTPUT_IMAGE_KEY
+from inference.core.workflows.core_steps.visualizations.base_colorable import (
+    ColorableVisualizationBlock,
+    ColorableVisualizationManifest,
 )
 from inference.core.workflows.entities.base import WorkflowImageData
 from inference.core.workflows.entities.types import (
@@ -17,8 +18,6 @@ from inference.core.workflows.entities.types import (
 )
 from inference.core.workflows.prototypes.block import BlockResult, WorkflowBlockManifest
 
-OUTPUT_IMAGE_KEY: str = "image"
-
 TYPE: str = "MaskVisualization"
 SHORT_DESCRIPTION = "Paints a mask over detected objects in an image."
 LONG_DESCRIPTION = """
@@ -28,7 +27,7 @@ from an instance segmentation to draw a mask using
 """
 
 
-class MaskManifest(VisualizationManifest):
+class MaskManifest(ColorableVisualizationManifest):
     type: Literal[f"{TYPE}"]
     model_config = ConfigDict(
         json_schema_extra={
@@ -55,7 +54,7 @@ class MaskManifest(VisualizationManifest):
     )
 
 
-class MaskVisualizationBlock(VisualizationBlock):
+class MaskVisualizationBlock(ColorableVisualizationBlock):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.annotatorCache = {}
